Fix payment method pills swapping
